refactor hero section
