+++ src/components/HeroSection.css
@@ -1,0 +1,29 @@
+.hero {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+    align-items: flex-start;
+  
+    background-color: #1f2937;
+    color: #f9faf8;
+  
+    padding: 200px 15%;
+  }
+  
+  .hero > h1 {
+      color: #f9faf8;
+      font-size: 48px;
+      font-weight: 900;
+      margin: 0;
+  }
+  
+  .hero > p {
+      color: #e5e7eb;
+      font-weight: 100;
+      font-size: 18px;
+      margin: 0;
+  }
+  
+  .hero-left {
+      width: 500px;
+  }
