+++ src/components/MainContent.css
@@ -1,34 +1,3 @@
-
-.hero {
-  display: flex;
-  width: 100%;
-  justify-content: space-between;
-  align-items: flex-start;
-
-  background-color: #1f2937;
-  color: #f9faf8;
-
-  padding: 200px 15%;
-}
-
-.hero h1 {
-    color: #f9faf8;
-    font-size: 48px;
-    font-weight: 900;
-    margin: 0;
-}
-
-.hero p {
-    color: #e5e7eb;
-    font-weight: 100;
-    font-size: 18px;
-    margin: 0;
-}
-
-.hero-left {
-    width: 500px;
-}
-
 .socials-buttons {
     display: flex;  
     align-items: center;
@@ -36,21 +5,6 @@
     margin: 16px 0;
     gap: 48px;
 }
-
-.socials-button {
-    padding: 0;
-    background-color: #1f2937;
-    border: 0;
-    flex: 0;
-    width: 32px;
-    height: 32px;
-}
-
-.socials-button img {
-    width: 32px;
-    height: 32px;
-}
-
 
 .image-slide {
     background-color: white;
