+++ src/components/buttons/MediumButtons.css
@@ -1,5 +1,4 @@
-.important-button {
-    border: none;
+.medium-button {
     padding: 8px 32px;
     border-radius: 8px;
 
@@ -15,6 +14,7 @@
 }
 
 .visit-blog {
+  border: none;
   background-color: #3882f6;
   color: #f9faf8;
 }
